requirements.txt
aiohttp
aria2p
async_generator
bs4
cowpy
# cryptg
cffi>=1.0.0
dnspython
emoji
gitpython
google-api-python-client
google-auth-oauthlib
googletrans>=2.4.0
google_images_download>=2.7.1
gTTS>=2.0.1
gTTS-token>=1.1.3
hachoir
httplib2
humanize
lxml
oauth2client
Pillow>=5.3.0
psycopg2
psycopg2-binary
pybase64>=0.4.0
pylast
pySmartDL
python-barcode
python-dotenv
telegraph
youtube-dl
pytz
qrcode
requests>=2.18.4
search-engine-parser>=0.4.2
speedtest-cli>=2.0.2
sqlalchemy>=1.2
telethon>=1.10.6
telethon-session-sqlalchemy>=0.2.6
urbandict>=0.5
wikipedia>=1.4.0
coffeehouse
